Refactor dark map handling to core
